+++ __docs__/wonder-blocks-dropdown/option-item.argtypes.tsx
@@ -1,4 +1,5 @@
 import * as React from "react";
+import {ArgTypes} from "@storybook/react";
 import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
 import Pill from "@khanacademy/wonder-blocks-pill";
 import {IconMappings} from "../wonder-blocks-icon/phosphor-icon.argtypes";
@@ -45,13 +46,13 @@ export default {
         control: {type: "boolean"},
         description: "Whether or not the option item is disabled.",
         table: {
-            defaultValue: {summary: false},
+            defaultValue: {summary: "false"},
             type: {summary: "boolean"},
         },
         type: {name: "boolean", required: true},
     },
     onClick: {
-        control: {type: null},
+        control: {type: undefined},
         description: `Optional user-supplied callback when this item is called.`,
         table: {
             type: {summary: "() => unknown"},
@@ -75,7 +76,11 @@ export default {
             defaultValue: {summary: "none"},
             type: {summary: `"none" | "inset" | "full-width"`},
         },
-        type: {name: `"none" | "inset" | "full-width"`, required: false},
+        type: {
+            name: "enum",
+            value: ["none", "inset", "full-width"],
+            required: false,
+        },
     },
     leftAccessory: {
         options: Object.keys(AccessoryMappings) as Array<React.ReactNode>,
@@ -85,7 +90,7 @@ export default {
         table: {
             type: {summary: "React.Node"},
         },
-        type: {name: "React.Node", required: false},
+        type: {name: "other", value: "React.Node", required: false},
     },
     rightAccessory: {
         options: Object.keys(AccessoryMappings) as Array<React.ReactNode>,
@@ -95,7 +100,7 @@ export default {
         table: {
             type: {summary: "React.Node"},
         },
-        type: {name: "React.Node", required: false},
+        type: {name: "other", value: "React.Node", required: false},
     },
     subtitle1: {
         control: {
@@ -121,4 +126,4 @@ export default {
             },
         },
     },
-};
+} satisfies ArgTypes;