+++ packages/wonder-blocks-testing/src/harness/adapters/render-state.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import {KindError, Errors} from "@khanacademy/wonder-stuff-core";
 import {RenderStateRoot} from "@khanacademy/wonder-blocks-core";
 
-import type {TestHarnessAdapter} from "../types";
+import type {TestHarnessAdapter} from "@khanacademy/wonder-blocks-testing-core";
 
 //
 type Config = true;
@@ -13,11 +13,15 @@ type Config = true;
 export const defaultConfig: Config | null = null;
 
 /**
- * Test harness adapter for supporting portals.
+ * Test harness adapter for supporting render state-based hooks and components.
  *
- * Some components rely on rendering with a React Portal. This adapter ensures
- * that the DOM contains a mounting point for the portal with the expected
- * identifier.
+ * Some components and hooks utilize the render state context to manage what
+ * they render and when. In order for this to work, a `RenderStateRoot`
+ * component must be present to track the current render state.
+ *
+ * This adapter wraps the children in a `RenderStateRoot` component to enable
+ * the render state context. This adapter should be used when testing components
+ * that rely on the render state.
  */
 export const adapter: TestHarnessAdapter<Config> = (
     children: React.ReactNode,