+++ packages/wonder-blocks-tokens/src/index.ts
@@ -1,18 +1,27 @@
+// primitive tokens
 import {border} from "./tokens/border";
 import {color} from "./tokens/color";
 import {font} from "./tokens/font";
 import {spacing} from "./tokens/spacing";
 
+// semantic tokens
+import {semanticColor} from "./tokens/semantic-color";
+
+// utils
 import {mix, fade} from "./util/utils";
 
 export {
     /**
-     * Core tokens for the Wonder Blocks design system.
+     * Primitive tokens for the Wonder Blocks design system.
      */
     border,
     color,
     font,
     spacing,
+    /**
+     * Semantic tokens.
+     */
+    semanticColor,
     /**
      * Utility functions for working with colors.
      */