+++ packages/wonder-blocks-i18n/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/wonder-blocks-i18n
 
+## 3.1.0
+
+### Minor Changes
+
+-   fc7dc5e2: Add loadTranslations/clearTranslations methods to wonder-blocks-i18n.
+
+## 3.0.1
+
+### Patch Changes
+
+-   559e82d5: Update to build tooling, generating smaller output
+
 ## 3.0.0
 
 ### Major Changes